requirements.txt
boto3==1.17.36
botocore==1.20.36
certifi==2020.12.5
chardet==4.0.0
dropbox==11.4.1
idna==2.10
jmespath==0.10.0
ply==3.11
python-dateutil==2.8.1
python-dotenv==0.17.0
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
stone==3.2.1
urllib3==1.26.4
